fix: druga tura bledow kompilacji (platform API: MediaSession/MediaController, MediaProjectionManager, ClipboardManager, HttpUrl/Headers okhttp, DiscoveredTables.value, performGlobalAction, PlaybackParams, getActiveNotifications, CalendarTools parseFlexible, AlarmTools nullable, NixiOrb this.size)

+++ app/src/main/java/dev/nixi/audio/AudioBus.kt
@@ -134,7 +134,7 @@ object AudioBus {
             val s = buf[i].toInt()
             sum += (s * s).toLong()
         }
-        return sqrt(sum / len) / 32768f
+        return (sqrt(sum.toDouble() / len) / 32768.0).toFloat()
     }
 }
